+++ common/src/main/java/earth/terrarium/adastra/common/systems/GravityApiImpl.java
@@ -3,6 +3,7 @@ package earth.terrarium.adastra.common.systems;
 import earth.terrarium.adastra.api.events.AdAstraEvents;
 import earth.terrarium.adastra.api.planets.PlanetApi;
 import earth.terrarium.adastra.api.systems.GravityApi;
+import earth.terrarium.adastra.common.config.AdAstraConfig;
 import earth.terrarium.adastra.common.constants.PlanetConstants;
 import earth.terrarium.adastra.common.handlers.PlanetHandler;
 import earth.terrarium.adastra.common.planets.Planet;
@@ -40,6 +41,7 @@ public class GravityApiImpl implements GravityApi {
 
     @Override
     public float getGravity(Entity entity) {
+        if (AdAstraConfig.disableGravity) return 1;
         float gravity = getGravity(entity.level(), BlockPos.containing(entity.getX(), entity.getEyeY(), entity.getZ()));
         return AdAstraEvents.EntityGravityEvent.post(entity, gravity);
     }
@@ -68,6 +70,7 @@ public class GravityApiImpl implements GravityApi {
 
     @Override
     public void entityTick(Level level, LivingEntity entity, Vec3 travelVector, BlockPos movementAffectingPos) {
+        if (AdAstraConfig.disableGravity) return;
         if (!entity.isControlledByLocalInstance()) return;
 
         boolean touchingSomething = false;