+++ common/src/main/java/earth/terrarium/adastra/common/systems/TemperatureApiImpl.java
@@ -3,6 +3,7 @@ package earth.terrarium.adastra.common.systems;
 import earth.terrarium.adastra.api.events.AdAstraEvents;
 import earth.terrarium.adastra.api.planets.PlanetApi;
 import earth.terrarium.adastra.api.systems.TemperatureApi;
+import earth.terrarium.adastra.common.config.AdAstraConfig;
 import earth.terrarium.adastra.common.constants.PlanetConstants;
 import earth.terrarium.adastra.common.handlers.PlanetHandler;
 import earth.terrarium.adastra.common.items.armor.SpaceSuitItem;
@@ -85,6 +86,7 @@ public class TemperatureApiImpl implements TemperatureApi {
 
     @Override
     public void entityTick(ServerLevel level, LivingEntity entity) {
+        if (AdAstraConfig.disableTemperature) return;
         if (entity.getType().is(ModEntityTypeTags.CAN_SURVIVE_IN_SPACE)) return;
         if (this.isHot(level, entity.blockPosition())) {
             if (entity.getType().is(ModEntityTypeTags.CAN_SURVIVE_EXTREME_HEAT)) return;